requirements.txt
gradio==4.19.2  # Pin to avoid bugs in later versions
langchain>=0.1.16
langchain-openai>=0.0.8
pypdf>=3.0.0
python-dotenv>=1.0.0
openai>=1.0.0
tiktoken==0.6.0 
pydantic==1.10.14  # Pin for FastAPI/Gradio compatibility
fastapi<0.110  # Pin for Pydantic v1 compatibility